Fruit Machine / Anchor Management Issues
Issue with Fruit Machine applies to Anchor Management as well. Updating README and comments in the script to reflect both and included attempted fix for Fruit Machine. If using a selector works, that's probably a better method than posting a duplicate form anyways, so we should just change everything to that if we get it to work.